Remove checks on 'finished' in animation completion blocks
When the in-call status bar is displayed, transitioning from a child
controller that displays the status bar to a child controller that
hides it (or vice versa) cancels the opening/closing animation.

This, together with the check on 'finished', was causing the calls
to 'didOpen' and 'didClose' to be skipped thus leaving the controller
in an incomplete state.